feat: Add HomePageExplore data
This commit adds the HomePageExplore data file, which contains an array of objects representing different course categories and their corresponding courses. Each course object includes information such as the course heading, description, level, and number of lessons. This addition enhances the functionality of the application by providing data for the Explore section on the homepage.